small fix (return type)

+++ features-chat/src/main/java/com/nullit/features_chat/persistance/dao/DialogDao.kt
@@ -13,6 +13,6 @@ interface DialogDao {
 
     @Transaction
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertDialogWithMembers(dialogEntity: DialogEntity): Int
+    suspend fun insertDialogWithMembers(dialogEntity: DialogEntity): Long
 
 }
